Release 2.14.0

+++ Mindbox/GuaranteedDeliveryManager/DeliveryOperation.swift
@@ -24,9 +24,15 @@ class DeliveryOperation: AsyncOperation, @unchecked Sendable {
     var onCompleted: ((_ event: Event, _ error: MindboxError?) -> Void)?
 
     override func main() {
+        guard !isCancelled else { finish(); return }
+        
         Logger.common(message: "[DeliveryOperation] Sending event `\(event.type.rawValue)` with transactionId: \(event.transactionId))", level: .info, category: .delivery)
+        
         eventRepository.send(event: event) { [weak self] result in
-            guard let self = self else { return }
+            guard let self else { return }
+            defer { self.finish() }
+            guard !self.isCancelled else { return }
+
             switch result {
             case .success:
                 self.onCompleted?(self.event, nil)
@@ -36,13 +42,15 @@ class DeliveryOperation: AsyncOperation, @unchecked Sendable {
                 self.onCompleted?(self.event, error)
                 Logger.common(message: "[DeliveryOperation] Did send event `\(event.type.rawValue)` with transactionId: \(self.event.transactionId) failed with error: \(error.localizedDescription)",
                               level: .error, category: .delivery)
+                
+                guard !self.isCancelled else { return }
+                
                 if case let MindboxError.protocolError(response) = error, HTTPURLResponseStatusCodeValidator(statusCode: response.httpStatusCode).isClientError {
                     try? self.databaseRepository.delete(event: self.event)
                 } else {
                     try? self.databaseRepository.update(event: self.event)
                 }
             }
-            self.finish()
         }
     }
 }
